+++ src/condor_utils/read_multiple_logs.h
@@ -210,7 +210,7 @@ public:
 protected:
 	friend class CheckEvents;
 
-	static unsigned int hashFuncJobID(const CondorID &key);
+	static size_t hashFuncJobID(const CondorID &key);
 
 private:
 	void cleanup();
